[~] Move IPOperations to separate project #2582
