+++ roles/lib/files/FWO.Report/ReportConnections.cs
@@ -1,10 +1,11 @@
-using FWO.GlobalConstants;
 using FWO.Api.Data;
 using FWO.Api.Client;
 using FWO.Report.Filter;
 using FWO.Config.Api;
 using System.Text;
 using FWO.Logging;
+using FWO.GlobalConstants;
+using FWO.GlobalFunctions;
 
 namespace FWO.Report
 {
@@ -95,7 +96,7 @@ namespace FWO.Report
                 report.AppendLine($"<td>{connection.Id}</td>");
                 if(isInterface)
                 {
-                    report.AppendLine($"<td>{GlobalFunc.ShowBool(connection.IsPublished)}</td>");
+                    report.AppendLine($"<td>{GlobalFuncs.ShowBool(connection.IsPublished)}</td>");
                 }
                 if(IsGlobalComSvc)
                 {
