+++ roles/ui/files/FWO.UI/Pages/Monitoring/MonitorImportStatus.razor
@@ -3,6 +3,7 @@
 @using FWO.Api.Data
 @using FWO.Api.Client
 @using FWO.Config.Api
+@using FWO.GlobalFunctions
 @using FWO.Ui.Shared
 
 @page "/monitoring/import_status"
@@ -70,7 +71,7 @@
                 <Template>
                     @if (context.LastImport != null && context.LastImport.Length > 0)
                     {
-                        @(GlobalFunc.ShowBool(context.LastImport[0].SuccessfulImport));
+                        @(GlobalFuncs.ShowBool(context.LastImport[0].SuccessfulImport));
                     }
                 </Template>
             </Column>
