+++ roles/ui/files/FWO.UI/Pages/Request/DisplayImplTaskTable.razor
@@ -2,6 +2,7 @@
 @using FWO.GlobalConstants
 @using FWO.Api.Data
 @using FWO.Config.Api
+@using FWO.GlobalFunctions
 @using FWO.Ui.Services
 
 @attribute [Authorize(Roles = $"{Roles.Admin}, {Roles.Requester}, {Roles.Approver}, {Roles.Planner}, {Roles.Implementer}, {Roles.Reviewer}, {Roles.Auditor}, {Roles.FwAdmin}")]
@@ -162,7 +163,7 @@
                         <Column TableItem="KeyValuePair<string, bool>" Title="@(userConfig.GetText("device"))" Field="@(x => x.Key)" Sortable="true" Filterable="true" />
                         <Column TableItem="KeyValuePair<string, bool>" Title="@(userConfig.GetText("impltask_created"))" Field="@(x => x.Value)" Sortable="false">
                             <Template>
-                                @(GlobalFunc.ShowBool(context.Value))
+                                @(GlobalFuncs.ShowBool(context.Value))
                             </Template>
                         </Column>
                     </Table>
