+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsCredentials.razor
@@ -5,6 +5,7 @@
 @using FWO.Api.Client.Queries
 @using FWO.Encryption
 @using FWO.Config.Api
+@using FWO.GlobalFunctions
 @using FWO.Middleware.Client
 
 @page "/settings/credentials"
@@ -63,7 +64,7 @@
         <Column TableItem="ImportCredential" Title="@(userConfig.GetText("is_key_pair"))" Field="@(x => x.IsKeyPair)"
             Sortable="true" Filterable="true">
             <Template>
-                @(GlobalFunc.ShowBool(context.IsKeyPair))
+                @(GlobalFuncs.ShowBool(context.IsKeyPair))
             </Template>
         </Column>
