+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsGroups.razor
@@ -3,6 +3,7 @@
 @using FWO.GlobalConstants
 @using FWO.Api.Data
 @using FWO.Api.Client
+@using FWO.GlobalFunctions
 @using FWO.Ui.Services
 @using FWO.Middleware.Client
 @using FWO.Config.Api
@@ -59,7 +60,7 @@
         <Column TableItem="UserGroup" Title="@(userConfig.GetText("name"))" Field="@(x => x.Name)" Sortable="true" Filterable="true" />
         <Column TableItem="UserGroup" Title="@(userConfig.GetText("owner_group"))" Field="@(x => x.OwnerGroup)" Sortable="true" Filterable="true">
             <Template>
-                @(GlobalFunc.ShowBool(context.OwnerGroup))
+                @(GlobalFuncs.ShowBool(context.OwnerGroup))
             </Template>
         </Column>
         <Column TableItem="UserGroup" Title="@(userConfig.GetText("users"))" Field="@(x => x.Users)" Sortable="false" Filterable="false">
@@ -90,7 +91,7 @@
                         }
                         else
                         {
-                            @(GlobalFunc.ShowBool(actGroup.OwnerGroup))
+                            @(GlobalFuncs.ShowBool(actGroup.OwnerGroup))
                         }
                     </div>
                 </div>
