+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsLdap.razor
@@ -1,5 +1,6 @@
 ﻿@using BlazorTable
 @using System.Net
+@using FWO.GlobalFunctions
 @using RestSharp
 
 @using FWO.GlobalConstants
@@ -40,7 +41,7 @@
         </Column>
         <Column TableItem="UiLdapConnection" Title="@(userConfig.GetText("active"))" Field="@(x => x.Active)" Sortable="true" Filterable="true">
             <Template>
-                @(GlobalFunc.ShowBool(context.Active))
+                @(GlobalFuncs.ShowBool(context.Active))
             </Template>
         </Column>
         <Column TableItem="UiLdapConnection" Title="@(userConfig.GetText("name"))" Field="@(x => x.Name)" Sortable="true" Filterable="true" />
@@ -56,22 +57,22 @@
         </Column>
         <Column TableItem="UiLdapConnection" Title="@(userConfig.GetText("search_user"))" Field="@(x => x.SearchUser)" Sortable="true">
             <Template>
-                @(GlobalFunc.ShowBool(context.SearchUser != null && context.SearchUser != ""))
+                @(GlobalFuncs.ShowBool(context.SearchUser != null && context.SearchUser != ""))
             </Template>
         </Column>
         <Column TableItem="UiLdapConnection" Title="@(userConfig.GetText("write_user"))" Field="@(x => x.WriteUser)" Sortable="true">
             <Template>
-                @(GlobalFunc.ShowBool(context.IsWritable()))
+                @(GlobalFuncs.ShowBool(context.IsWritable()))
             </Template>
         </Column>
         <Column TableItem="UiLdapConnection" Title="@(userConfig.GetText("role_handling"))" Field="@(x => x.RoleSearchPath)" Sortable="true">
             <Template>
-                @(GlobalFunc.ShowBool(context.HasRoleHandling()))
+                @(GlobalFuncs.ShowBool(context.HasRoleHandling()))
             </Template>
         </Column>
         <Column TableItem="UiLdapConnection" Title="@(userConfig.GetText("group_handling"))" Field="@(x => (x.GroupSearchPath))" Sortable="true">
             <Template>
-                @(GlobalFunc.ShowBool(context.HasGroupHandling()))
+                @(GlobalFuncs.ShowBool(context.HasGroupHandling()))
             </Template>
         </Column>
         <Column TableItem="UiLdapConnection" Title="@(userConfig.GetText("tenant"))" Field="@(x => x.TenantId)" Sortable="true">
