+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsManagements.razor
@@ -5,6 +5,7 @@
 @using FWO.Api.Client.Queries
 @using FWO.Config.Api
 @using FWO.DeviceAutoDiscovery
+@using FWO.GlobalFunctions
 @using FWO.Middleware.Client
 
 @page "/settings/managements"
@@ -92,7 +93,7 @@
         <Column TableItem="Management" Title="@(userConfig.GetText("importer_host"))" Field="@(x => x.ImporterHostname)" Sortable="true" Filterable="true" />
         <Column TableItem="Management" Title="@(userConfig.GetText("import_enabled"))" Field="@(x => x.ImportDisabled)" Sortable="true">
             <Template>
-                @(GlobalFunc.ShowBool(!context.ImportDisabled))
+                @(GlobalFuncs.ShowBool(!context.ImportDisabled))
             </Template>
         </Column>
         <Column TableItem="Management" Title="@(userConfig.GetText("debug_level"))" Field="@(x => x.DebugLevel)" Sortable="true" Filterable="true" />
