+++ roles/ui/files/FWO.UI/Pages/Settings/SettingsUsers.razor
@@ -4,6 +4,7 @@
 @using FWO.Api.Data
 @using FWO.Api.Client
 @using FWO.Api.Client.Queries
+@using FWO.GlobalFunctions
 @using FWO.Middleware.Client
 @using FWO.Config.Api
 @using FWO.Middleware.RequestParameters
@@ -94,7 +95,7 @@
         <Column TableItem="UiUser" Title="@(userConfig.GetText("pwd_chg_req"))" Field="@(x => x.PasswordMustBeChanged)" Sortable="true" Filterable="true" >
             <Template>
                 <div data-toggle="tooltip" title="@(userConfig.PureLine("H5278"))">
-                    @(GlobalFunc.ShowBool(context.PasswordMustBeChanged))
+                    @(GlobalFuncs.ShowBool(context.PasswordMustBeChanged))
                 </div>
             </Template>
         </Column>
